refactor(swap-service): move logger-only helpers into swaps/utils.ts
Introduce a module-level Logger('SwapsUtils') in utils.ts and move helpers
whose only service dependency was this.logger. Keeps the service as an
orchestration interface and co-locates related logic in utils.

Moved:
- fetchUsdPrices
- resolveSwapSellAmountUsd

getReferralCode and resolveAffiliateAddress stay on the service (prisma/
userServiceClient dependencies would be awkward to pass as params).
