+++ apps/swap-service/src/swaps/utils.ts
@@ -1,11 +1,18 @@
+import { Logger } from '@nestjs/common'
 import type { Swap as PrismaSwap } from '@prisma/client'
 
+import { CreateSwapDto } from '@shapeshift/shared-types'
+import { baseUnitToPrecision } from '@shapeshift/shared-utils'
 import { bnOrZero } from '@shapeshiftoss/chain-adapters'
 import type { SwapperSpecificMetadata } from '@shapeshiftoss/swapper'
 import type { Asset } from '@shapeshiftoss/types'
 
 import { getSwapperFeeStrategy } from '../utils/affiliateFeeAsset'
-import type { Swap } from './types'
+import { getAssetPriceUsd } from '../utils/pricing'
+
+import type { Swap, UsdPrices } from './types'
+
+const logger = new Logger('SwapsUtils')
 
 export const toSwap = (swap: PrismaSwap): Swap => ({
   ...swap,
@@ -14,9 +21,11 @@ export const toSwap = (swap: PrismaSwap): Swap => ({
   metadata: swap.metadata as SwapperSpecificMetadata,
 })
 
+// formatAmount up to 8 decimals, no trailing zeros
 export const formatAmount = (amount: string | number): string => {
-  // up to 8 decimals, no trailing zeros
-  return bnOrZero(amount).toFixed(8).replace(/\.?0+$/, '')
+  return bnOrZero(amount)
+    .toFixed(8)
+    .replace(/\.?0+$/, '')
 }
 
 export const getAffiliateCommissionRate = (
@@ -24,10 +33,8 @@ export const getAffiliateCommissionRate = (
   verifiedBps: number,
   shapeshiftBps: number,
 ): number => {
-  if (origin === 'web') {
-    // Referrer earns shapeshiftBps of volume (shapeshiftBps / verifiedBps of the fee).
-    return shapeshiftBps / verifiedBps
-  }
+  // Referrer earns shapeshiftBps of volume (shapeshiftBps / verifiedBps of the fee).
+  if (origin === 'web') return shapeshiftBps / verifiedBps
   if (!origin || verifiedBps <= shapeshiftBps) return 0
   return (verifiedBps - shapeshiftBps) / verifiedBps
 }
@@ -52,6 +59,42 @@ export const estimateAffiliateFeeAmount = (
   }
 }
 
+export const fetchUsdPrices = async (
+  data: CreateSwapDto,
+  affiliateFeeAssetId: string | null,
+): Promise<UsdPrices> => {
+  try {
+    const [sellAssetUsd, buyAssetUsd, affiliateAssetUsd] = await Promise.all([
+      getAssetPriceUsd(data.sellAsset.assetId),
+      getAssetPriceUsd(data.buyAsset.assetId),
+      affiliateFeeAssetId ? getAssetPriceUsd(affiliateFeeAssetId) : Promise.resolve<number | null>(null),
+    ])
+
+    const sellAmountUsd = sellAssetUsd
+      ? bnOrZero(baseUnitToPrecision(data.sellAmountCryptoBaseUnit, data.sellAsset.precision))
+          .times(sellAssetUsd)
+          .toFixed(2)
+      : null
+
+    return {
+      sellAmountUsd,
+      buyAssetUsd: buyAssetUsd?.toString() ?? null,
+      affiliateAssetUsd: affiliateAssetUsd?.toString() ?? null,
+    }
+  } catch (err) {
+    logger.warn(`Failed to fetch USD prices for swap ${data.swapId}:`, err)
+    return { sellAmountUsd: null, buyAssetUsd: null, affiliateAssetUsd: null }
+  }
+}
+
+export const resolveSwapSellAmountUsd = (swap: { swapId: string; sellAmountUsd: string | null }): number | null => {
+  if (!swap.sellAmountUsd) {
+    logger.warn(`Missing sellAmountUsd for swap ${swap.swapId}, skipping`)
+    return null
+  }
+  return parseFloat(swap.sellAmountUsd)
+}
+
 export const resolveFeeAssetPrice = (swap: {
   sellAsset: unknown
   buyAsset: unknown
@@ -62,15 +105,17 @@ export const resolveFeeAssetPrice = (swap: {
   affiliateFeeAssetId: string | null
 }): string | null => {
   if (!swap.affiliateFeeAssetId) return null
+
   const sellAssetObj = swap.sellAsset as Asset
   const buyAssetObj = swap.buyAsset as Asset
+
   if (swap.affiliateFeeAssetId === sellAssetObj.assetId && swap.sellAmountUsd) {
     // Back-derive sell price from stored USD value.
     const sellAmountPrecision = bnOrZero(swap.sellAmountCryptoBaseUnit).div(bnOrZero(10).pow(sellAssetObj.precision))
     return sellAmountPrecision.isZero() ? null : bnOrZero(swap.sellAmountUsd).div(sellAmountPrecision).toFixed()
   }
-  if (swap.affiliateFeeAssetId === buyAssetObj.assetId) {
-    return swap.buyAssetUsd
-  }
+
+  if (swap.affiliateFeeAssetId === buyAssetObj.assetId) return swap.buyAssetUsd
+
   return swap.affiliateAssetUsd
 }
